removed useLoginForm
